Comply with PEP8 style guide

+++ backend/app/models.py
@@ -1,16 +1,18 @@
 from datetime import datetime, timezone
 from enum import Enum
 
-from pydantic import computed_field, BaseModel
+from pydantic import BaseModel
 from sqlmodel import Field, SQLModel, Relationship
 
 
 def get_datetime_utc() -> datetime:
     return datetime.now(timezone.utc)
 
+
 class OrderType(Enum):
     BUY = "BUY"
     SELL = "SELL"
+
 
 class Order(BaseModel):
     stock_id: int
@@ -18,21 +20,24 @@ class Order(BaseModel):
     quantity: int
     order_type: OrderType
 
+
 class UserStockLink(SQLModel, table=True):
     user_id: int | None = Field(default=None, foreign_key="user.id", primary_key=True)
     stock_id: int = Field(default=None, foreign_key="stock.id", primary_key=True)
+
 
 class User(SQLModel, table=True):
     id: int | None = Field(default=None, primary_key=True)
     username: str = Field(index=True)
     email: str = Field(index=True)
     password: str = Field()
-    balance: int = Field(default=0)
+    balance: float = Field(default=0)
     disabled: bool = Field(default=False)
 
     transactions: list["Transaction"] = Relationship(back_populates="user")
     positions: list["Position"] = Relationship(back_populates="user")
     watched_stocks: list["Stock"] = Relationship(back_populates="users", link_model=UserStockLink)
+
 
 class Stock(SQLModel, table=True):
     id: int | None = Field(default=None, primary_key=True)
@@ -41,6 +46,7 @@ class Stock(SQLModel, table=True):
     exchange: str = Field(index=True)
 
     users: list["User"] = Relationship(back_populates="watched_stocks", link_model=UserStockLink)
+
 
 class Transaction(SQLModel, table=True):
     id: int | None = Field(default=None, primary_key=True)
@@ -54,10 +60,11 @@ class Transaction(SQLModel, table=True):
     user_id: int = Field(default=None, foreign_key="user.id")
     user: User | None = Relationship(back_populates="transactions")
 
+
 class Position(SQLModel, table=True):
     id: int | None = Field(default=None, primary_key=True)
     quantity: int = Field(default=1)
-    average_price: int = Field(default=0)
+    average_price: float = Field(default=0)
     type: OrderType = Field(default=OrderType.BUY)
 
     stock_id: int = Field(default=None, foreign_key="stock.id")
@@ -65,9 +72,11 @@ class Position(SQLModel, table=True):
     user_id: int = Field(default=None, foreign_key="user.id")
     user: User | None = Relationship(back_populates="positions")
 
+
 class Token(BaseModel):
     access_token: str
     token_type: str
+
 
 class TokenData(BaseModel):
     username: str | None = None
